# docs/swagger.yaml
definitions:
...
  pb.APIResponse:
    properties:
      code:
        type: integer
      response:
        type: string
    type: object
...
  pb.FutureDetail:
    properties:
      category:
        type: string
      code:
        type: string
      currency:
        type: string
      day_trade:
        type: string
      delivery_date:
        type: string
      delivery_month:
        type: string
      exchange:
        type: string
      limit_down:
        type: number
      limit_up:
        type: number
      margin_trading_balance:
        type: integer
      multiplier:
        type: integer
      name:
        type: string
      option_right:
        type: string
      reference:
        type: number
      security_type:
        type: string
      short_selling_balance:
        type: integer
      strike_price:
        type: integer
      symbol:
        type: string
      target_code:
        type: string
      underlying_code:
        type: string
      underlying_kind:
        type: string
      unit:
        type: integer
      update_date:
        type: string
    type: object
  pb.FutureDetailList:
    properties:
      list:
        items:
          $ref: '#/definitions/pb.FutureDetail'
        type: array
    type: object
...
  pb.HistoryKbar:
    properties:
      close:
        type: number
      code:
        type: string
      high:
        type: number
      kbar_time:
        $ref: '#/definitions/timestamppb.Timestamp'
      low:
        type: number
      open:
        type: number
      volume:
        type: integer
    type: object
  pb.HistoryKbarList:
    properties:
      list:
        items:
          $ref: '#/definitions/pb.HistoryKbar'
        type: array
    type: object
...
  pb.OptionDetail:
    properties:
      category:
        type: string
      code:
        type: string
      currency:
        type: string
      day_trade:
        type: string
      delivery_date:
        type: string
      delivery_month:
        type: string
      exchange:
        type: string
      limit_down:
        type: number
      limit_up:
        type: number
      margin_trading_balance:
        type: integer
      multiplier:
        type: integer
      name:
        type: string
      option_right:
        type: string
      reference:
        type: number
      security_type:
        type: string
      short_selling_balance:
        type: integer
      strike_price:
        type: number
      symbol:
        type: string
      target_code:
        type: string
      underlying_code:
        type: string
      underlying_kind:
        type: string
      unit:
        type: integer
      update_date:
        type: string
    type: object
  pb.OptionDetailList:
    properties:
      list:
        items:
          $ref: '#/definitions/pb.OptionDetail'
        type: array
    type: object
...
  pb.StockDetail:
    properties:
      category:
        type: string
      code:
        type: string
      currency:
        type: string
      day_trade:
        type: string
      delivery_date:
        type: string
      delivery_month:
        type: string
      exchange:
        type: string
      limit_down:
        type: number
      limit_up:
        type: number
      margin_trading_balance:
        type: integer
      multiplier:
        type: integer
      name:
        type: string
      option_right:
        type: string
      reference:
        type: number
      security_type:
        type: string
      short_selling_balance:
        type: integer
      strike_price:
        type: integer
      symbol:
        type: string
      target_code:
        type: string
      underlying_code:
        type: string
      underlying_kind:
        type: string
      unit:
        type: integer
      update_date:
        type: string
    type: object
  pb.StockDetailList:
    properties:
      list:
        items:
          $ref: '#/definitions/pb.StockDetail'
        type: array
    type: object
...
  structpb.ListValue:
    properties:
      values:
        description: Repeated field of dynamically typed values.
        items:
          $ref: '#/definitions/structpb.Value'
        type: array
    type: object
  structpb.Value:
    properties:
      kind:
        description: "The kind of value.\n\nTypes that are valid to be assigned to
          Kind:\n\n\t*Value_NullValue\n\t*Value_NumberValue\n\t*Value_StringValue\n\t*Value_BoolValue\n\t*Value_StructValue\n\t*Value_ListValue"
    type: object
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
info:
  contact: {}
  description: Capitan V1 Srv's API docs
  title: Capitan V1 OpenAPI
  version: v0.0
paths:
  /api/capitan/v1/basic/future/kbar:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.HistoryKbarList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pb.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pb.APIResponse'
      security:
      - JWT: []
      summary: Get futures
      tags:
      - Basic V1
  /api/capitan/v1/basic/futures:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.FutureDetailList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pb.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pb.APIResponse'
      security:
      - JWT: []
      summary: Get futures
      tags:
      - Basic V1
  /api/capitan/v1/basic/options:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.OptionDetailList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pb.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pb.APIResponse'
      security:
      - JWT: []
      summary: Get options
      tags:
      - Basic V1
  /api/capitan/v1/basic/stocks:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.StockDetailList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pb.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pb.APIResponse'
      security:
      - JWT: []
      summary: Get stocks
      tags:
      - Basic V1
...
  /api/capitan/v1/stream/subscribe/codes:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structpb.ListValue'
      security:
      - JWT: []
      summary: Get all trade records
      tags:
      - Stream V1
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"